AddMigration/CreateDb

+++ Models/DTO/Airtime.cs
@@ -8,7 +8,8 @@ namespace Amos.VTUCORE3._1.Models.DTO
     public class Airtime
     {
         public int Id { get; set; }
-        public double Discount { get; set; }
+        public decimal Discount { get; set; }
+
         public Service Service { get; set; }
 
     }
